component  managed
